third_party/blink/web_tests/platform/linux-chrome/external/wpt/web-animations/timing-model/timelines/timelines-expected.txt

This is a testharness.js-based test.
[PASS] Timeline time increases once per animation frame
[FAIL] Timeline time increases once per animation frame in an iframe
  assert_greater_than: Timeline time within an iframe increases between animation frames expected a number greater than 0 but got 0
[PASS] Timeline time should be the same for all RAF callbacks in an animation frame
[PASS] Performs a microtask checkpoint after updating timelines
[PASS] Runs finished promise before animation events
Harness: the test ran to completion.
